requirements.txt
# Python Version: 3.10.14
blinker==1.7.0
chardet==5.2.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.3
img2pdf==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
packaging==24.0
pdf2image==1.17.0
pikepdf==8.15.0
pillow==10.3.0
PyPDF2==3.0.1
reportlab==4.1.0
Werkzeug==3.0.2
wrapt==1.16.0
